requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==19.3.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.4
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.5
cloud-init==22.4.2
command-not-found==0.3
configobj==5.0.6
cryptography==36.0.2
DateTime==4.3
dbus-python==1.2.16
defusedxml==0.7.1
distro-info===0.23ubuntu1
Django==4.1.5
django-ckeditor-5==0.2.2
django-cleanup==8.0.0
django-filter==21.1
fail2ban==0.11.1
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.2
incuna-mail==4.1.1
incuna-pigeon==0.1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
language-selector==0.1
libretranslatepy==2.1.1
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==1.1.0
more-itertools==4.2.0
multidict==6.0.4
netifaces==0.10.4
oauthlib==3.2.0
openai==0.27.8
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
pyrsistent==0.15.5
pyserial==3.4
pyTelegramBotAPI==4.12.0
python-apt==2.0.1
python-dotenv==0.20.0
python3-openid==3.2.0
pytils==0.4.1
pytz==2021.3
PyYAML==5.3.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.4.1
sqlparse==0.4.2
supervisor==4.1.0
telebot==0.0.5
tqdm==4.65.0
translate==3.6.1
tzdata==2022.7
ubuntu-advantage-tools==8001
ufw==0.36
urllib3==1.26.9
webencodings==0.5.1
yarl==1.9.2
zipp==3.7.0
zope.interface==5.4.0
